requirements.txt
# Core
python-dotenv
requests

# Google Sheets (Service Account auth)
gspread
google-auth

# AI / Email generation
anthropic

# Data processing
pandas
